+++ linphone-desktop/src/components/camera/CameraPreview.cpp
@@ -71,8 +71,8 @@ QOpenGLFramebufferObject *CameraPreviewRenderer::createFramebufferObject (const 
   // It's not the same thread as render.
   coreManager->lockVideoRender();
 
-  mContextInfo->width = size.width();
-  mContextInfo->height = size.height();
+  mContextInfo->width = static_cast<GLuint>(size.width());
+  mContextInfo->height = static_cast<GLuint>(size.height());
   mContextInfo->functions = MSFunctions::getInstance()->getFunctions();
   mUpdateContextInfo = true;
